It displays totals properly now.
FossilOrigin-Name: 4fadcc401c4d7e5bb66d384c0170105d6fc97243d5d0507148f51e34c42dffa6

+++ Common/Data.cs
@@ -111,7 +111,7 @@ namespace RV.InvNew.Common
         [JsonInclude] public double uselling;
         [JsonInclude] public double umarked;
         [JsonInclude] public double quantity;
-        [JsonInclude] public double VatCategory;
+        [JsonInclude] public long VatCategory;
         [JsonInclude] public double VatPercent;
         [JsonInclude] public double VatAmount;
         [JsonInclude] public double Total;
